\name{binary}
\alias{binary}
\docType{data}
\title{Example data set (binary endpoint)}
\description{
Example data set based on a trial with a binary endpoint.
}
\usage{data(binary)}
\format{
  A data set with 359 observations and 11 variables:
  \describe{
    \item{\code{outcome}}{Binary outcome variable (0 or 1). The value of 1 indicates the beneficial outcome.}
    \item{\code{treatment}}{Binary treatment variable (0 or 1). The values of 0 and 1 correspond to the control and experimental treatments, respectively.}
    \item{\code{cont1}}{Continuous covariate used in the primary analysis model.}
    \item{\code{cont2}}{Continuous covariate used in the primary analysis model.}
    \item{\code{class1}}{Class/categorical covariate used in the primary analysis model.}
    \item{\code{class2}}{Class/categorical covariate used in the primary analysis model.}  
    \item{\code{biomarker1}}{Numeric biomarker.}  
    \item{\code{biomarker2}}{Numeric biomarker.}  
    \item{\code{biomarker3}}{Numeric biomarker.}  
    \item{\code{biomarker4}}{Numeric biomarker.}  
    \item{\code{biomarker5}}{Nominal biomarker.}  
    }
}
\keyword{datasets}
